add state machines display

+++ backend/api/src/main/kotlin/de/keepmealive3d/adapters/data/StateMachine.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class StateData(
     val id: String,
-    val posX: Long,
-    val posY: Long,
+    val posX: Int,
+    val posY: Int,
     val connectedTo: MutableList<String>,
 )
